+++ src/query/sql/src/planner/binder/bind_table_reference/bind_join.rs
@@ -33,8 +33,8 @@ use crate::normalize_identifier;
 use crate::optimizer::ir::ColumnSet;
 use crate::optimizer::ir::RelExpr;
 use crate::optimizer::ir::SExpr;
-use crate::optimizer::operator::FlattenInfo;
-use crate::optimizer::operator::SubqueryRewriter;
+use crate::optimizer::optimizers::operator::FlattenInfo;
+use crate::optimizer::optimizers::operator::SubqueryRewriter;
 use crate::optimizer::OptimizerContext;
 use crate::planner::binder::scalar::ScalarBinder;
 use crate::planner::binder::Binder;
