+++ src/query/sql/src/planner/optimizer/optimizer.rs
@@ -28,15 +28,15 @@ use crate::optimizer::distributed::BroadcastToShuffleOptimizer;
 use crate::optimizer::distributed::SortAndLimitPushDownOptimizer;
 use crate::optimizer::ir::Memo;
 use crate::optimizer::ir::SExpr;
-use crate::optimizer::operator::DeduplicateJoinConditionOptimizer;
-use crate::optimizer::operator::PullUpFilterOptimizer;
-use crate::optimizer::operator::RuleNormalizeAggregateOptimizer;
-use crate::optimizer::operator::RuleStatsAggregateOptimizer;
-use crate::optimizer::operator::SingleToInnerOptimizer;
-use crate::optimizer::operator::SubqueryRewriter;
+use crate::optimizer::optimizers::operator::DeduplicateJoinConditionOptimizer;
+use crate::optimizer::optimizers::operator::PullUpFilterOptimizer;
+use crate::optimizer::optimizers::operator::RuleNormalizeAggregateOptimizer;
+use crate::optimizer::optimizers::operator::RuleStatsAggregateOptimizer;
+use crate::optimizer::optimizers::operator::SingleToInnerOptimizer;
+use crate::optimizer::optimizers::operator::SubqueryRewriter;
+use crate::optimizer::optimizers::recursive::RecursiveOptimizer;
 use crate::optimizer::optimizers::CascadesOptimizer;
 use crate::optimizer::optimizers::DPhpy;
-use crate::optimizer::optimizers::RecursiveOptimizer;
 use crate::optimizer::statistics::CollectStatisticsOptimizer;
 use crate::optimizer::util::contains_local_table_scan;
 use crate::optimizer::util::contains_warehouse_table_scan;
