+++ src/query/sql/src/planner/optimizer/optimizers/operator/aggregate/mod.rs
