+++ src/query/sql/src/planner/optimizer/optimizers/operator/aggregate/normalize_aggregate.rs
