+++ src/query/sql/src/planner/optimizer/optimizers/operator/aggregate/stats_aggregate.rs
