+++ src/query/sql/src/planner/optimizer/optimizers/operator/decorrelate/decorrelate.rs
@@ -34,9 +34,9 @@ use crate::optimizer::ir::ColumnSet;
 use crate::optimizer::ir::Matcher;
 use crate::optimizer::ir::RelExpr;
 use crate::optimizer::ir::SExpr;
-use crate::optimizer::operator::FlattenInfo;
-use crate::optimizer::operator::SubqueryRewriter;
-use crate::optimizer::operator::UnnestResult;
+use crate::optimizer::optimizers::operator::FlattenInfo;
+use crate::optimizer::optimizers::operator::SubqueryRewriter;
+use crate::optimizer::optimizers::operator::UnnestResult;
 use crate::plans::BoundColumnRef;
 use crate::plans::CastExpr;
 use crate::plans::ComparisonOp;
