+++ src/query/sql/src/planner/optimizer/optimizers/operator/decorrelate/flatten_plan.rs
@@ -28,8 +28,8 @@ use crate::binder::Visibility;
 use crate::optimizer::ir::ColumnSet;
 use crate::optimizer::ir::RelExpr;
 use crate::optimizer::ir::SExpr;
-use crate::optimizer::operator::FlattenInfo;
-use crate::optimizer::operator::SubqueryRewriter;
+use crate::optimizer::optimizers::operator::FlattenInfo;
+use crate::optimizer::optimizers::operator::SubqueryRewriter;
 use crate::plans::Aggregate;
 use crate::plans::AggregateFunction;
 use crate::plans::AggregateMode;
