+++ src/query/sql/src/planner/optimizer/optimizers/operator/join/single_to_inner.rs
