+++ src/query/sql/src/planner/optimizer/optimizers/recursive/recursive.rs
