+++ src/query/sql/src/planner/optimizer/rule/rewrite/rule_push_down_filter_join.rs
@@ -20,8 +20,8 @@ use crate::binder::JoinPredicate;
 use crate::optimizer::ir::Matcher;
 use crate::optimizer::ir::RelExpr;
 use crate::optimizer::ir::SExpr;
-use crate::optimizer::operator::InferFilterOptimizer;
-use crate::optimizer::operator::JoinProperty;
+use crate::optimizer::optimizers::operator::InferFilterOptimizer;
+use crate::optimizer::optimizers::operator::JoinProperty;
 use crate::optimizer::rule::constant::false_constant;
 use crate::optimizer::rule::constant::is_falsy;
 use crate::optimizer::rule::rewrite::push_down_filter_join::can_filter_null;
