+++ src/query/sql/tests/optimizer/filter/infer_filter_test.rs
@@ -18,7 +18,7 @@ use databend_common_expression::types::DataType;
 use databend_common_expression::types::NumberDataType;
 use databend_common_expression::types::NumberScalar;
 use databend_common_expression::types::Scalar;
-use databend_common_sql::optimizer::operator::InferFilterOptimizer;
+use databend_common_sql::optimizer::optimizers::operator::InferFilterOptimizer;
 use databend_common_sql::planner::binder::ColumnBinding;
 use databend_common_sql::planner::binder::Visibility;
 use databend_common_sql::planner::plans::BoundColumnRef;
